\name{ChessGmooG}
\alias{ChessGmooG-package}
\docType{package}
\title{
ChessGmooG: Chess datasets analysed in Chapter 8 of "Getting (more out of) Graphics"
}

\description{
The two datasets of chess ratings from December 2020 and December 2015 are each bigger than 1Mb, so they are provided in this package.
}

\details{
Most of the datasets analysed in the book are available in GmooG and other R packages.
}

\author{
Antony Unwin unwin@math.uni-augsburg.de

}


